src/App/requirements.txt
streamlit
Pillow
numpy
ultralytics
opencv-python-headless